requirements.txt
adlfs==2024.7.0
aiobotocore==2.15.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
azure-core==1.32.0
azure-datalake-store==0.0.53
azure-identity==1.19.0
azure-storage-blob==12.24.0
bcrypt==4.2.0
botocore==1.35.36
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
croniter==5.0.1
cryptography==43.0.1
dataclasses-json==0.5.9
decorator==5.1.1
diskcache==5.6.3
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
flyteidl==1.13.7
flytekit==1.13.13
frozenlist==1.5.0
fsspec==2024.10.0
gcsfs==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpickle==4.0.0
keyring==25.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
marshmallow-enum==1.5.1
marshmallow-jsonschema==0.13.0
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.5.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==2.1.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.2
passlib==1.7.4
portalocker==2.10.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.27.5
protoc-gen-openapiv2==0.0.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.6.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytimeparse==1.1.8
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rich-click==1.8.4
rsa==4.9
s3fs==2024.10.0
sentry-sdk==2.18.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.38.5
statsd==4.0.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
union==0.1.97
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
wheel==0.43.0
wrapt==1.16.0
yarl==1.17.2
zipp==3.21.0
